+++ lib/eve_online.rb
@@ -73,7 +73,7 @@ require 'eve_online/esi/character_skills'
 
 require 'eve_online/esi/server_status'
 
-require 'eve_online/esi/ancestries'
+require 'eve_online/esi/universe_ancestries'
 require 'eve_online/esi/universe_asteroid_belt'
 require 'eve_online/esi/universe_bloodlines'
 require 'eve_online/esi/universe_categories'