+++ lib/eve_online/esi/universe_ancestries.rb
@@ -2,7 +2,7 @@
 
 module EveOnline
   module ESI
-    class Ancestries < Base
+    class UniverseAncestries < Base
       API_ENDPOINT = 'https://esi.evetech.net/v1/universe/ancestries/?datasource=%<datasource>s'
 
       def ancestries