+++ spec/eve_online/esi/universe_ancestries_spec.rb
@@ -2,7 +2,7 @@
 
 require 'spec_helper'
 
-describe EveOnline::ESI::Ancestries do
+describe EveOnline::ESI::UniverseAncestries do
   specify { expect(subject).to be_a(EveOnline::ESI::Base) }
 
   specify { expect(described_class::API_ENDPOINT).to eq('https://esi.evetech.net/v1/universe/ancestries/?datasource=%<datasource>s') }